feat(app-bot): route inference through user's bound free MiniMax bot (validated) + drop job/callback
Owner-chosen + prod-validated: each install binds a free official bot (slot 1); gateway injects
a server-side persona pre-prompt and relays via client-speak; bot replies toUser; app polls
/api/client/pending. Removes app_bot_job/callback/result. Persona + quota stay server-side.
Epic card_6d0f2746, card_12e6d33b.
