+++ backend/app-bots-config.js
@@ -31,31 +31,17 @@ const APP_BOTS = {
 };
 
 // ============================================
-// Inference entity (option B — card_6d0f2746 / card_12e6d33b)
+// Inference model (owner-chosen + prod-validated — card_6d0f2746 / card_12e6d33b)
 //
 // Replies are NOT generated by a server-side LLM call (prod has no
-// ANTHROPIC_API_KEY → 502). Instead the gateway dispatches each job to an
-// always-on EClaw bot ENTITY running on an external device (async). That entity
-// reads the job envelope, generates a reply, and POSTs it back to
-// /api/app-bot/chat/callback. This block identifies that entity; values come
-// from env with safe fallbacks so the file stays free of committed secrets.
+// ANTHROPIC_API_KEY → 502) and NOT by a custom async inference entity. Instead,
+// each app install binds a FREE official MiniMax bot to its OWN device (slot 1)
+// during onboarding. The gateway injects a server-side persona pre-prompt and
+// relays the user's message to that bound bot via the normal client-speak
+// delivery path; the bot replies `toUser` and the app reads it via
+// GET /api/client/pending. No custom job/callback envelope is needed — the
+// persona and quota stay entirely server-side.
 // ============================================
-const APP_BOT_INFERENCE = {
-    deviceId: process.env.APP_BOT_INFERENCE_DEVICE_ID || null,
-    entityId: process.env.APP_BOT_INFERENCE_ENTITY_ID ? Number(process.env.APP_BOT_INFERENCE_ENTITY_ID) : null,
-    botSecret: process.env.APP_BOT_INFERENCE_BOT_SECRET || null,
-    publicCode: process.env.APP_BOT_INFERENCE_PUBLIC_CODE || null,
-};
-
-/**
- * Whether an inference entity is configured. The gateway must return 503
- * (inference_unavailable) when this is false — dispatch resolves the target by
- * publicCode, so publicCode is the operative gate.
- * @returns {boolean}
- */
-function isInferenceConfigured() {
-    return typeof APP_BOT_INFERENCE.publicCode === 'string' && APP_BOT_INFERENCE.publicCode.length > 0;
-}
 
 /**
  * Resolve an (appId, personaId) pair to its server-side config.
@@ -78,6 +64,4 @@ function getPersona(appId, personaId) {
 module.exports = {
     APP_BOTS,
     getPersona,
-    APP_BOT_INFERENCE,
-    isInferenceConfigured
 };
